refactor(scan): prepare to introduce Custom support
Make it clearer what in VxScan is generic and what is specifically Plustek.
