+++ apps/scan/backend/src/app.ts
@@ -34,14 +34,14 @@ import {
   exportCastVoteRecordsToUsbDrive,
 } from './cvrs/export';
 import { PrecinctScannerInterpreter } from './interpret';
-import { PrecinctScannerStateMachine } from './state_machine';
-import { Workspace } from './util/workspace';
-import { Usb } from './util/usb';
 import {
+  PrecinctScannerStateMachine,
   ConfigurationError,
   PrecinctScannerConfig,
   PrecinctScannerStatus,
 } from './types';
+import { Workspace } from './util/workspace';
+import { Usb } from './util/usb';
 import { getMachineConfig } from './machine_config';
 import { CVR_EXPORT_FORMAT } from './globals';
 import { DefaultMarkThresholds } from './store';
