+++ apps/scan/backend/src/server.ts
@@ -4,7 +4,7 @@ import { LogEventId, Logger, LogSource } from '@votingworks/logging';
 import { buildApp } from './app';
 import { PORT } from './globals';
 import { PrecinctScannerInterpreter } from './interpret';
-import { PrecinctScannerStateMachine } from './state_machine';
+import { PrecinctScannerStateMachine } from './types';
 import { Usb } from './util/usb';
 import { Workspace } from './util/workspace';
